~0 artifacts

+++ CPSC591-Project/SSAOFirstPassFramebuffer.cpp
@@ -27,7 +27,7 @@ void SSAOFirstPassFramebuffer::create() {
 	//Position buffer
 	glGenTextures(1, &positionTex);
 	glBindTexture(GL_TEXTURE_2D, positionTex);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, SceneParameters::getScreenWidth(), SceneParameters::getScreenHeight(), 0, GL_RGBA, GL_FLOAT, NULL);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, SceneParameters::getScreenWidth(), SceneParameters::getScreenHeight(), 0, GL_RGBA, GL_FLOAT, NULL);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
